reset database where appropriate
