+++ Dockerfile
@@ -14,14 +14,7 @@ COPY [".", "/home/neuro/bidsMReye"]
 RUN pip install --upgrade pip && \
     pip3 install -e .
 
-RUN bidsmreye_model && \
-    bidsmreye_model --model_name 1_guided_fixations && \
-    bidsmreye_model --model_name 2_pursuit && \
-    bidsmreye_model --model_name 3_openclosed && \
-    bidsmreye_model --model_name 3_pursuit && \
-    bidsmreye_model --model_name 4_pursuit && \
-    bidsmreye_model --model_name 5_free_viewing && \
-    bidsmreye_model --model_name 1to5
+RUN bidsmreye_model
 
 ENTRYPOINT [ "//home/neuro/bidsMReye/entrypoint.sh" ]
 COPY ["./docker/entrypoint.sh", \
