+++ Makefile
@@ -150,7 +150,6 @@ generalize: ## demo: predicts labels of MOAE dataset
 				participant \
 				--action generalize \
 				-vv \
-				--reset_database \
 				--non_linear_coreg
 
 
@@ -174,6 +173,7 @@ ds002799_prepare: get_ds002799
 				--action prepare \
 				--participant_label 302 307 \
 				--space MNI152NLin2009cAsym \
+				--reset_database \
 				--run 1 2
 
 
@@ -195,6 +195,7 @@ ds002799: clean-ds002799 get_ds002799
 				--participant_label 302 307 \
 				--space MNI152NLin2009cAsym \
 				--run 1 2 \
+				--reset_database \
 				-vv
 
 ## DOCKER
@@ -232,7 +233,7 @@ docker_generalize:
 				--action generalize
 
 docker_ds002799: get_ds002799
-	datalad unlock $$PWD/tests/data/ds002799/derivatives/fmriprep/sub-30[27]/ses-*/func/*run-*preproc*bold*
+# datalad unlock $$PWD/tests/data/ds002799/derivatives/fmriprep/sub-30[27]/ses-*/func/*run-*preproc*bold*
 	docker run --rm -it \
 				-v $$PWD/tests/data/ds002799/derivatives/fmriprep:/home/neuro/data \
 				-v $$PWD/outputs/ds002799/derivatives:/home/neuro/outputs/ \
@@ -244,4 +245,5 @@ docker_ds002799: get_ds002799
 				--participant_label 302 307 \
 				--space MNI152NLin2009cAsym \
 				--run 1 2 \
+				--reset_database \
 				-vv
